% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_events.R
\name{add_events}
\alias{add_events}
\title{Title}
\usage{
add_events(
  df = NA,
  x = NA,
  y = NA,
  xend = NA,
  yend = NA,
  heatmap = FALSE,
  shot = FALSE,
  lines = FALSE,
  bgCol = "white",
  shirt = NA,
  eventArgs = list(),
  shotArgs = list(),
  heatmapArgs = list(),
  lineArgs = list()
)
}
\arguments{
\item{lineArgs}{}
}
\description{
Title
}
