% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart}
\alias{bar_chart}
\title{bar_chart}
\usage{
bar_chart(
  df,
  x,
  y,
  id,
  z,
  positive = c(TRUE, FALSE),
  fillCol = NA,
  highCol = NA
)
}
\arguments{
\item{df}{Data frame containing KPI and Label, x and y}

\item{x}{Column containing KPI}

\item{y}{Label}

\item{id}{Id to highlight}

\item{z}{Selected KPI - from shiny input}

\item{positive}{True if higher is better for KPI}

\item{fillCol}{Color of non-highlighted ids}

\item{highCol}{Highlighted id color}
}
\value{
A bar plot, chosen id is highlighted, and mean is added as a red line - ordering is done by "positive"
}
\description{
bar_chart
}
\examples{
no example
}
