% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_order.R
\name{comp_order}
\alias{comp_order}
\title{Competition Order}
\usage{
comp_order(competitionId)
}
\arguments{
\item{competitionId}{Wyscout competition ID.}
}
\value{
An integer rank indicating the order.
}
\description{
Orders divisions by age and division level.
This function is designed for Wyscout competition IDs.
}
\examples{

A <- data.frame(competitionId = c(355, 328, 100))
A$Order <- comp_order(A$competitionId)

}
