% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_chart.R
\name{create_beautiful_radarchart}
\alias{create_beautiful_radarchart}
\title{radar chart}
\usage{
create_beautiful_radarchart(
  data,
  color = NA,
  axisCol = NA,
  vlabels = colnames(data),
  vlcex = 0.7,
  caxislabels = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{data}{Data}

\item{color}{Radar color}

\item{axisCol}{Axis color}

\item{vlabels}{Axis labels}

\item{vlcex}{Axis size}

\item{caxislabels}{Caxis labels}

\item{title}{Title}

\item{...}{}
}
\value{
A costum version of the fmsb radar chart
}
\description{
radar chart
}
