% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_chart_single.R
\name{create_radarchart_single}
\alias{create_radarchart_single}
\title{radar_chart_single}
\usage{
create_radarchart_single(
  df = NA,
  x = NA,
  y = NA,
  lab = NA,
  col = NA,
  title = "Put title here",
  subTitle = "Put subtitle here",
  provider = "Put provider here"
)
}
\arguments{
\item{df}{Dataframe with, x, y and labels}

\item{x}{Variable names}

\item{y}{Percentile values}

\item{lab}{Data labels}

\item{col}{Color}

\item{title}{Title}

\item{subTitle}{Subtitle}

\item{provider}{Data provider}
}
\value{
Returns a radar plot for a single team/player
}
\description{
Plots a radar chart for a single team/player
}
