% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_col.R
\name{div_col}
\alias{div_col}
\title{div_col}
\usage{
div_col(type = NA, color = NA)
}
\arguments{
\item{type}{String indicating type for default colors}

\item{color}{Custom color input - either a color from colors() or a color in hex code}
}
\value{
Color
}
\description{
Given either type or custom input checks if the input is a color and returns it for plotting use
}
\examples{
div_col(color = "white")
div_col(type = "reference")
}
