% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_chart.R
\name{gauge}
\alias{gauge}
\title{gauge}
\usage{
gauge(
  value,
  KPI,
  teams = 12,
  textCol = NA,
  bestCol = NA,
  secCol = NA,
  thrCol = NA,
  worstCol = NA
)
}
\arguments{
\item{value}{Percentile value * 100}

\item{KPI}{Plot title}

\item{teams}{Number of teams in league (default: 12).}

\item{textCol}{Text color}

\item{bestCol}{Best quantile color}

\item{secCol}{Second quantile color}

\item{thrCol}{Third quantile color}

\item{worstCol}{Fourth quantile color}
}
\value{
Gauge plot with text based on 12 team league
}
\description{
Creates gauge chart for comparison across teams
}
\examples{
gauge(50, "Test")
gauge(50, "Test", worstCol = "white", bestCol = "steelblue", secCol = "yellow", thrCol = "lightblue")
}
