% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goal_mouth.R
\name{goal_mouth}
\alias{goal_mouth}
\title{goal_mouth}
\usage{
goal_mouth(
  team = input$sel_team_opp,
  season = input$sel_season,
  start = input$gw_prep_def_slider[1],
  end = input$gw_prep_def_slider[2],
  df,
  z,
  y,
  type,
  goalCol = NA,
  saveCol = NA,
  provider = "OPTA",
  keeperDetails = c(TRUE, FALSE),
  teams,
  name
)
}
\arguments{
\item{team}{Selected team - from input}

\item{season}{Selected season - from input}

\item{start}{Start of period - from input}

\item{end}{End of period - from input}

\item{df}{Data with shots}

\item{z}{Column with goal mouth z coordinate}

\item{y}{Column with goal mouth y coordinate}

\item{type}{Column containing event types}

\item{goalCol}{Point color for goals}

\item{saveCol}{Point color for saves}

\item{provider}{Provider for caption}

\item{keeperDetails}{Binary - if true adds keeper name to plot}

\item{teams}{Column containing team names}

\item{name}{Column containing keeper names}
}
\value{
Adds shots to the goal_frame plot - highlighting goals
}
\description{
goal_mouth
}
