% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not In Operator (\%nin\%)}
\usage{
a \%nin\% b
}
\arguments{
\item{a}{Object to search for.}

\item{b}{Object to search in.}
}
\value{
TRUE if 'a' is not in 'b', else FALSE.
}
\description{
Checks if an object is not in another object.
}
\examples{
a <- 1
b <- c(1, 3, 5)
a \%nin\% b
}
