% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_sparkline_vb.R
\name{hc_theme_sparkline_vb}
\alias{hc_theme_sparkline_vb}
\title{hc_theme_sparkline_vb}
\usage{
hc_theme_sparkline_vb(...)
}
\arguments{
\item{...}{Additional parameters for further customizing the Highcharter theme.}
}
\value{
A customized Highcharter theme for sparkline visualizations.
}
\description{
Customize the Highcharter theme for sparkline visualizations.
}
