% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_to_head.R
\name{head_to_head}
\alias{head_to_head}
\title{head_to_head}
\usage{
head_to_head(
  df,
  home = NA,
  textCol = NA,
  provider = "OPTA",
  title = "HEAD TO HEAD"
)
}
\arguments{
\item{df}{Data frame}

\item{home}{Name of home team}

\item{textCol}{Color used for text}

\item{provider}{Data provider}

\item{title}{Title - defaults to HEAD TO HEAD}
}
\value{
A head-to-head chart with data from a match
}
\description{
head_to_head
}
