% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_hexcolor.R
\name{is_hexcolor}
\alias{is_hexcolor}
\title{is_hexcolor}
\usage{
is_hexcolor(text = NA)
}
\arguments{
\item{text}{String to test for hex color code.}
}
\value{
TRUE if valid hex color code, FALSE otherwise.
}
\description{
Check if the input text is a valid hex color code.
}
\examples{
is_hexcolor("#ffffff")
}
