% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart_con.R
\name{line_chart_con}
\alias{line_chart_con}
\title{Line Chart on Continuous X Axis}
\usage{
line_chart_con(
  df,
  x,
  y,
  labels,
  ref = NA,
  trend = NA,
  lCol = NA,
  KPI = "Put KPI here",
  Periods = "Put x label here",
  Title = NA,
  Subtitle = NA,
  provider = "Put provider here",
  refCol = NA
)
}
\arguments{
\item{df}{Data frame.}

\item{x}{Column with data for the x axis.}

\item{y}{Column with data for the y axis.}

\item{labels}{Column with point labels.}

\item{ref}{Reference value on the y axis.}

\item{trend}{Set to "TRUE" to add smoothed trend.}

\item{lCol}{Line color.}

\item{KPI}{Y axis label.}

\item{Periods}{X axis label.}

\item{Title}{Title of the plot.}

\item{Subtitle}{Subtitle of the plot.}

\item{provider}{Data provider.}

\item{refCol}{Color of the reference line.}
}
\value{
A line chart on a continuous x axis.
}
\description{
Creates a line chart on a continuous x axis.
}
