% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart_phys.R
\name{line_chart_phys}
\alias{line_chart_phys}
\title{Line Chart for Physical Performance}
\usage{
line_chart_phys(
  df,
  date = "Date",
  parameter = input$sel_phys_par,
  name = "TeamName",
  team = NA,
  match = NA,
  compare = input$sel_phys_overlay,
  debug = 0
)
}
\arguments{
\item{df}{A data frame containing the physical performance data.}

\item{date}{Column name for the date variable (default: "Date").}

\item{parameter}{Selected physical parameter (default: input$sel_phys_par).}

\item{name}{Column name for team names (default: "TeamName").}

\item{team}{Selected team (default: NA).}

\item{match}{Selected match (default: NA).}

\item{compare}{Selected teams for comparison (default: input$sel_phys_overlay).}

\item{debug}{Debugging flag (0 or 1, default: 0).}
}
\value{
A line chart for visualizing physical performance data.
}
\description{
Generate a line chart to visualize physical performance data.
}
\examples{
line_chart_phys(df = phys_data, parameter = "Total Distance (km)", team = "Vejle Boldklub", compare = "FC Midtjylland")
}
