% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(x, type = c("Z", "O"))
}
\arguments{
\item{x}{Series of values or a data frame column}

\item{type}{Choose between Z-score or 0 to 1 normalization}
}
\value{
Series of values scaled between 0 and 1
}
\description{
Normalizes a series of values to be between 0 and 1
}
\examples{
normalize(c(0, 5, 10), type="Z")

normalize(c(0, 5, 10), type="O") * 100
}
