% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_dev_lines.R
\name{player_dev_lines}
\alias{player_dev_lines}
\title{Player Development Lines}
\usage{
player_dev_lines(
  df,
  competitionId = "competitionId",
  season = "season",
  value = "value",
  comp = "comp",
  KPI = "Put KPI here",
  provider = "Put provider here"
)
}
\arguments{
\item{df}{Data frame with player data.}

\item{competitionId}{Column name of Wyscout competitionId. (Default: "competitionId")}

\item{season}{Column name of x axis. (Default: "season")}

\item{value}{Column name of y axis. (Default: "value")}

\item{comp}{Column name of competition name for facet_wrap. (Default: "comp")}

\item{KPI}{Name of y axis KPI. (Default: "Put KPI here")}

\item{provider}{Name of data provider. (Default: "Put provider here")}
}
\value{
Plot of KPI by season - facet_wrap by competition - ordered by age and level.
}
\description{
Creates a plot of a specific KPI by season, facet-wrapped by competition, and ordered by age and level.
}
