% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_chart.R
\name{point_chart}
\alias{point_chart}
\title{point_chart}
\usage{
point_chart(df, value, facet, playerCol = NA, otherCol = NA)
}
\arguments{
\item{df}{Data frame containing players, the value to compare and facet variable}

\item{value}{The value to compare in percentile format}

\item{facet}{Variable to use for facet wrap}

\item{playerCol}{Highlighted player color}

\item{otherCol}{Color for other players}
}
\value{
Point chart used in the scouting module
}
\description{
point_chart
}
