% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_chart.R
\name{probability_chart}
\alias{probability_chart}
\title{probablity chart}
\usage{
probability_chart(
  wp,
  barCol = NA,
  textCol = NA,
  accuracy = 1,
  provider = "OPTA"
)
}
\arguments{
\item{wp}{Data from win_prob reactive}

\item{barCol}{Color of bars}

\item{textCol}{Color of text}

\item{accuracy}{Decimals for label}

\item{provider}{Provider for caption}
}
\value{
Win probability chart
}
\description{
probablity chart
}
