% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarchart_div.R
\name{radarchart_div}
\alias{radarchart_div}
\title{Radar Chart}
\usage{
radarchart_div(
  df = NA,
  x = NA,
  y = NA,
  lab = NA,
  group = NA,
  col = NA,
  title = "TITLE",
  subTitle = "SUBTITLE",
  provider = "Opta"
)
}
\arguments{
\item{df}{A data frame containing the data for the radar chart.}

\item{x}{A character vector of variable names representing categories or attributes.}

\item{y}{A numeric vector of percentile values for each category.}

\item{lab}{Data labels to display on the chart (only displayed for the first group). (Default: NA)}

\item{group}{A grouping variable (e.g., players or teams). (Default: NA)}

\item{col}{A character vector of colors (up to 4) for the radar chart. (Default: NA)}

\item{title}{A character string for the title of the radar chart. (Default: "TITLE")}

\item{subTitle}{A character string for the subtitle of the radar chart. (Default: "SUBTITLE")}

\item{provider}{A character string indicating the data provider. (Default: "Opta")}
}
\value{
A radar chart visualizing data for a single team/player.
}
\description{
Plots a radar chart to visualize data across multiple categories for up to 4 groups.
}
