% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_win_prob.R
\name{sim_win_prob}
\alias{sim_win_prob}
\title{sim_win_prob}
\usage{
sim_win_prob(home, away, m = 1e+05)
}
\arguments{
\item{home}{A vector representing the home team's shots.}

\item{away}{A vector representing the away team's shots.}

\item{m}{Number of simulations. Default is 100000.}
}
\value{
A numeric vector with probabilities of home, draw, and away outcomes.
}
\description{
Calculate the win probabilities of home, draw, and away outcomes for a match
using a Monte Carlo simulation strategy.
}
\examples{
sim_win_prob(c(0.6, 0.8), c(0.45, 0.06))
}
