% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbox.R
\name{vbox}
\alias{vbox}
\title{vbox}
\usage{
vbox(
  value,
  title,
  sparkobj = NULL,
  subtitle,
  info = NULL,
  icon = NULL,
  color = "aqua",
  width = 4,
  href = NULL
)
}
\arguments{
\item{value}{Value}

\item{title}{Title}

\item{sparkobj}{Spark object}

\item{subtitle}{Subtitle}

\item{info}{Info}

\item{icon}{Icon}

\item{color}{Color}

\item{width}{Width}

\item{href}{Href}
}
\description{
vbox
}
