% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_events.R
\name{add_events}
\alias{add_events}
\title{add_events}
\usage{
add_events(
  df = NA,
  x = NA,
  y = NA,
  xend = NA,
  yend = NA,
  heatmap = FALSE,
  shot = FALSE,
  lines = FALSE,
  bgCol = "white",
  shirt = NA,
  eventArgs = list(),
  shotArgs = list(),
  heatmapArgs = list(),
  lineArgs = list()
)
}
\arguments{
\item{df}{Data frame}

\item{x}{X coordinate}

\item{y}{Y coordinate}

\item{xend}{X end coordinate}

\item{yend}{Y end coordinate}

\item{heatmap}{Set to "TRUE" to add heatmap based on x,y}

\item{shot}{Set to "TRUE" to switch to shot plot}

\item{lines}{Set to "TRUE" to add lines from x,y to xend,yend}

\item{bgCol}{Background color - set to shiny theme}

\item{shirt}{Column with shirt numbers - will be added to events}

\item{eventArgs}{Event arguments, usable: color and outcome (0 or 1)}

\item{shotArgs}{Shot arguments, usable: color and outcome ("goal" and "other")}

\item{heatmapArgs}{Heat map arguments: usable alpha, color and fill}

\item{lineArgs}{Lines arguments: usable linetype and color}
}
\value{
Returns set of ggplot layers to add to a ggplot
}
\description{
Add events to pitch plot - does not plot the pitch itself
}
