% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_events.R
\name{add_events}
\alias{add_events}
\title{Add Events to Pitch Plot}
\usage{
add_events(
  df = NA,
  x = NA,
  y = NA,
  xend = NA,
  yend = NA,
  provider = NA,
  size = 4,
  heatmap = FALSE,
  shot = FALSE,
  corners = FALSE,
  lines = FALSE,
  passZones = FALSE,
  goal_kick_helper = FALSE,
  pmCol = "black",
  bgCol = "white",
  textCol = "black",
  title = "Put title here",
  explanation = "Put explanation here",
  shirt = NA,
  eventArgs = list(),
  shotArgs = list(),
  cornerArgs = list(),
  heatmapArgs = list(),
  lineArgs = list()
)
}
\arguments{
\item{df}{A data frame containing event data.}

\item{x}{Column with X coordinates.}

\item{y}{Column with Y coordinates.}

\item{xend}{Column with X end coordinates.}

\item{yend}{Column with Y end coordinates.}

\item{provider}{Name of the data provider.}

\item{size}{Point size - defaults to 4, minimum is 3.}

\item{heatmap}{Set to TRUE to add a heatmap based on x, y.}

\item{shot}{Set to TRUE to switch to a shot plot.}

\item{corners}{Set to TRUE to switch to a corner plot.}

\item{lines}{Set to TRUE to add lines from x, y to xend, yend.}

\item{passZones}{Set to TRUE to add counts for passes to each zone.}

\item{goal_kick_helper}{Set to TRUE in goal_kick_plots while useing passZones.}

\item{pmCol}{Pitch marking colors for heatmap use.}

\item{bgCol}{Background color - set according to the Shiny theme.}

\item{textCol}{Text color.}

\item{title}{Title of the plot.}

\item{explanation}{Explanation of the chart.}

\item{shirt}{Column with shirt numbers - will be added to events.}

\item{eventArgs}{Event arguments, including color and outcome column (0 or 1).}

\item{shotArgs}{Shot arguments, including color and outcome column ("goal" or "other").}

\item{cornerArgs}{Use for Opta tags - arguments: colors and type column name.}

\item{heatmapArgs}{Heat map arguments, including alpha, color, fill, and type ("start" or "end").}

\item{lineArgs}{Line arguments, including linetype and color.}
}
\value{
A set of ggplot layers to add to a ggplot.
}
\description{
Adds events to a pitch plot without plotting the pitch itself.
}
\examples{
x <- 100
df <- data.frame(
  x = runif(x, 20, 75),
  y = runif(x, 10, 90),
  xend = runif(x, 40, 99),
  yend = runif(x, 10, 90),
  outcome = round(runif(x), 0)
)


p <- ggplot() +
  annotate_pitch() +
  add_events(
    df = df, x = "x", y = "y", xend = "xend", yend = "yend",
    provider = "Wyscout",
    lines = TRUE, passZones = FALSE, textCol = "black", bgCol = "white",
    corners = FALSE, cornerArgs = list(),
    eventArgs = list(outcome = "outcome"),
    shot = FALSE, shotArgs = list(),
    heatmap = TRUE, heatmapArgs = list(type = "end", outcome = "outcome")
  ) +
  theme_pitch()

p

}
