% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart}
\alias{bar_chart}
\title{Bar Chart}
\usage{
bar_chart(
  df,
  x,
  y,
  id = NA,
  KPI = "Name of KPI",
  negative = FALSE,
  median = FALSE,
  fillCol = NA,
  highCol = NA,
  refCol = NA
)
}
\arguments{
\item{df}{A data frame containing KPI and label information.}

\item{x}{Column containing the KPI values.}

\item{y}{Column containing the labels.}

\item{id}{ID of the team to be highlighted (default: NA, no team highlighted).}

\item{KPI}{The selected KPI (from Shiny input) - default: "Name of KPI".}

\item{negative}{Set to TRUE if lower values are better for the KPI (default: FALSE).}

\item{median}{Use median instead of mean for the reference line (default: FALSE).}

\item{fillCol}{Color of the bars for non-highlighted IDs (default: NA).}

\item{highCol}{Color for the bars of the highlighted ID (default: NA).}

\item{refCol}{Color of the reference line (default: NA).}
}
\value{
A bar plot where the chosen ID is highlighted, and the mean (or median) is indicated with a red line.
The ordering of bars is determined by the "negative" parameter.
}
\description{
Creates a bar chart for comparing multiple teams, highlighting a specific team with an optional reference line.
}
