% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart}
\alias{bar_chart}
\title{bar_chart}
\usage{
bar_chart(
  df,
  x,
  y,
  id,
  KPI,
  positive = c(TRUE, FALSE),
  median = c(TRUE, FALSE),
  fillCol = NA,
  highCol = NA,
  refCol = NA
)
}
\arguments{
\item{df}{Data frame containing KPI and Label, x and y}

\item{x}{Column containing KPI}

\item{y}{Label}

\item{id}{Id to highlight}

\item{KPI}{Selected KPI - from shiny input}

\item{positive}{True if higher is better for KPI}

\item{median}{Use median instead of mean}

\item{fillCol}{Color of non-highlighted ids}

\item{highCol}{Highlighted id color}

\item{refCol}{Color of reference line}
}
\value{
A bar plot, chosen id is highlighted, and mean is added as a red line - ordering is done by "positive"
}
\description{
Bar chart for comparison across multiple teams - id team will be highlighted
Reference line with either mean or median is added
Bar chart can be reordered using positiive=false if "lower is better"
}
