% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_order.R
\name{comp_order}
\alias{comp_order}
\title{comp_order}
\usage{
comp_order(competitionId)
}
\arguments{
\item{competitionId}{Wyscout competitionId}
}
\value{
Integer rank
}
\description{
Orders divisions age and division level
Works for Wyscout competition Id's
}
\examples{

A <- data.frame(competitionId = c(355, 328, 100))
A$Order <- comp_order(A$competitionId)

}
