% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_chart_single.R
\name{create_radarchart_single}
\alias{create_radarchart_single}
\title{create_radarchart_single}
\usage{
create_radarchart_single(
  df = NA,
  x = NA,
  y = NA,
  lab = NA,
  col = NA,
  title = "Put title here",
  subTitle = "Put subtitle here",
  provider = "Put provider here"
)
}
\arguments{
\item{df}{Dataframe containing the data for the radar chart.}

\item{x}{Variable names (e.g., categories or attributes).}

\item{y}{Percentile values for each category.}

\item{lab}{Data labels to display on the chart.}

\item{col}{Color to fill the bars (optional).}

\item{title}{Title for the radar chart.}

\item{subTitle}{Subtitle for the radar chart.}

\item{provider}{Data provider for the data source.}
}
\value{
Returns a radar plot for a single team/player.
}
\description{
Plots a radar chart for a single team/player.
}
