% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_transform.R
\name{def_transform}
\alias{def_transform}
\title{def_transform}
\usage{
def_transform(x, y)
}
\arguments{
\item{x}{data}
}
\value{
x subtracted from y
}
\description{
Subtracts x from y
}
\examples{

def_transform(10)
}
