% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_col.R
\name{div_col}
\alias{div_col}
\title{Div Color helper}
\usage{
div_col(type = NA, color = NA)
}
\arguments{
\item{type}{String indicating the type for default colors.}

\item{color}{Custom color input - either a color name from colors() or a hex color code.}
}
\value{
A color suitable for plotting.
}
\description{
Determines a color for plotting based on either a default type or a custom input.
}
\examples{
div_col(color = "white")
div_col(type = "reference")
}
