% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goal_frame.R
\name{goal_frame}
\alias{goal_frame}
\title{goal_frame}
\usage{
goal_frame(lowR = NA, lowM = NA, lowL = NA, upR = NA, upM = NA, upL = NA)
}
\value{
A ggplot with a goal frame
}
\description{
Defines a goal frame in ggplot for further use, needs to be printed for the plot to show
}
\examples{
print(goal_frame())
}
