% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goal_mouth.R
\name{goal_mouth}
\alias{goal_mouth}
\title{goal_mouth}
\usage{
goal_mouth(
  team = input$sel_team_opp,
  season = input$sel_season,
  start = input$gw_prep_def_slider[1],
  end = input$gw_prep_def_slider[2],
  df,
  z = "gk_z",
  y = "gk_y",
  type = "Event_type",
  goalCol = NA,
  saveCol = NA,
  provider = "OPTA",
  keeperDetails = TRUE,
  teams = "Conceding team",
  name = "Conceding keeper"
)
}
\arguments{
\item{team}{Selected team - from input variable as default}

\item{season}{Selected season - from from input variable as default}

\item{start}{Start of period - from from input variable as default}

\item{end}{End of period - from from input variable as default}

\item{df}{Data with shots on target}

\item{z}{Column with goal mouth z coordinate}

\item{y}{Column with goal mouth y coordinate}

\item{type}{Column containing event types ("Goal" is used, other types are discounted)}

\item{goalCol}{Point color for goals}

\item{saveCol}{Point color for saves}

\item{provider}{Provider for caption}

\item{keeperDetails}{Binary - if true adds keeper name to plot}

\item{teams}{Column containing team names}

\item{name}{Column containing keeper names}
}
\value{
Adds shots to the goal_frame plot - highlighting goals
}
\description{
goal_mouth
}
