% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_to_head.R
\name{head_to_head}
\alias{head_to_head}
\title{Head-to-Head Chart}
\usage{
head_to_head(
  df,
  x = "variable",
  y = "share",
  value = "value",
  names = "OfficialName",
  home = NA,
  textCol = NA,
  provider = "Opta",
  title = "HEAD TO HEAD"
)
}
\arguments{
\item{df}{Data frame containing the data for the head-to-head chart.}

\item{x}{Column with variables (e.g., match statistics or metrics).}

\item{y}{Column with shares (e.g., share of possession or goals).}

\item{value}{Column with numeric values (e.g., actual values of the metrics).}

\item{names}{Column with team names.}

\item{home}{Name of the home team (used for orientation).}

\item{textCol}{Color used for text labels (optional).}

\item{provider}{Data provider (default: "OPTA").}

\item{title}{Title of the chart (default: "HEAD TO HEAD").}
}
\value{
A head-to-head chart comparing the given data for two teams in a match.
}
\description{
Creates a head-to-head chart comparing the given data for two teams in a match.
}
