% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_hexcolor.R
\name{is_hexcolor}
\alias{is_hexcolor}
\title{is_hexcolor}
\usage{
is_hexcolor(text = NA)
}
\arguments{
\item{text}{String to test}
}
\value{
True is hexcode, false if not
}
\description{
is_hexcolor
}
\examples{
is_hexcolor("#ffffff")
}
