% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart_con.R
\name{line_chart_con}
\alias{line_chart_con}
\title{line_chart_con}
\usage{
line_chart_con(
  df,
  x,
  y,
  labels,
  ref = NA,
  trend = NA,
  lCol = NA,
  KPI = "Put KPI here",
  Periods = "Put x label here",
  Title = NA,
  Subtitle = NA,
  provider = "Put provider here",
  refCol = NA
)
}
\arguments{
\item{df}{Data frame}

\item{x}{Column with data for x axis}

\item{y}{Column with data for y axis}

\item{ref}{Reference value on y axis}

\item{trend}{Set to "TRUE" to add smoothed trend}

\item{lCol}{Line color}

\item{KPI}{Y axis label}

\item{Periods}{X axis label}

\item{Title}{Title of plot}

\item{Subtitle}{Subtitle of plot}

\item{provider}{Data provider}

\item{refCol}{Color of reference line}

\item{opponent}{Column with point labels}
}
\value{
Line chart on continuous x axis
}
\description{
Line chart on a continuous x axis
}
