% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart_con.R
\name{line_chart_con}
\alias{line_chart_con}
\title{line_chart_con}
\usage{
line_chart_con(
  df,
  x,
  y,
  idCl,
  id,
  ref = NA,
  KPI,
  Periods,
  Title = NA,
  Subtitle = NA,
  Caption = "Data by OPTA",
  refCol = NA
)
}
\arguments{
\item{df}{Data frame}

\item{x}{Column with data for x axis}

\item{y}{Column with data for y axis}

\item{idCl}{Column with data containing ids}

\item{id}{Id to show}

\item{ref}{Reference value on y axis}

\item{KPI}{Y axis label}

\item{Periods}{X axis label}

\item{Title}{Title of plot}

\item{Subtitle}{Subtitle of plot}

\item{Caption}{Caption for data provider}

\item{refCol}{Color of reference line}
}
\value{
Line chart on continuous x axis
}
\description{
line_chart_con
}
