% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{Series of values or a data frame column}
}
\value{
Series of values scaled between 0 and 1
}
\description{
normalize
}
\examples{
normalize(c(0, 5, 10))

normalize(c(0, 5, 10)) * 100
}
