% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_cbind.R
\name{normalize_cbind}
\alias{normalize_cbind}
\title{normalize}
\usage{
normalize_cbind(df, x, type = NA)
}
\arguments{
\item{df}{Data frame to work on and return}

\item{x}{Column to use for normalization}

\item{type}{Choose between Z-score ("Z") or 0 to 1 normalization ("1")}
}
\value{
Series of values scaled between 0 and 1
}
\description{
Normalizes a series of values to be between 0 and 1
}
\examples{

df <- data.frame(
  OfficialName = c("A", "B", "C", "D"),
  mean = runif(4)
)

df <- normalize_cbind(df, x = df$mean, type = "1")

df <- normalize_cbind(df, x = df$mean, type = "Z")
}
