% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_dev_lines.R
\name{player_dev_lines}
\alias{player_dev_lines}
\title{player_dev_lines}
\usage{
player_dev_lines(
  player_df,
  competitionId = "competitionId",
  season = "season",
  value = "value",
  comp = "comp",
  KPI = "Put KPI here",
  provider = "Put provider here"
)
}
\arguments{
\item{player_df}{Data frame with player data}

\item{competitionId}{Coloum name of Wyscout competitionId}

\item{season}{Coloum name of x axis}

\item{value}{Coloum name of y axis}

\item{comp}{Coloum name of competition name for facet_wrap}

\item{KPI}{Name of y axis KPI}

\item{provider}{Name of data provider}
}
\value{
Plot of KPI by season - facet_wrap by competition - order by age and level
}
\description{
player_dev_lines
}
