% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_chart.R
\name{point_chart}
\alias{point_chart}
\title{Point Chart}
\usage{
point_chart(
  df,
  name = "FullName",
  value = "value",
  facet = "variable",
  replaced = "TEST",
  playerCol = NA,
  repCol = NA,
  otherCol = NA,
  debug = 0
)
}
\arguments{
\item{df}{Data frame containing players, the value to compare, and the facet variable.}

\item{name}{Column with player names. (Default: "FullName")}

\item{value}{The value to compare in percentile format. (Default: "value")}

\item{facet}{Variable to use for facet wrap. (Default: "variable")}

\item{replaced}{Value in "name" column to highlight as replaced player. (Default: "TEST")}

\item{playerCol}{Highlighted player color.}

\item{repCol}{Color for player to replace.}

\item{otherCol}{Color for other players.}

\item{debug}{Debug mode. (Default: 0)}
}
\value{
Point chart used in the scouting module - depends on row_click() being defined.
}
\description{
Creates a point chart for comparing players' values using facet wrap for visualization.
}
