% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_win_prob.R
\name{poisson_win_prob}
\alias{poisson_win_prob}
\title{poisson_win_prob}
\usage{
poisson_win_prob(home_xg, away_xg)
}
\arguments{
\item{home_xg}{xG of home team}

\item{away_xg}{xG of away team}
}
\value{
A set of probabilities HomeWin\%,Draw\%,AwayWin\%
}
\description{
Calculates win probability between home and away using dpois()
}
\examples{
poisson_win_prob(home_xg=1.2,away_xg=1)
}
