% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarchart_div.R
\name{radarchart_div}
\alias{radarchart_div}
\title{radarchart_div}
\usage{
radarchart_div(
  df = NA,
  x = NA,
  y = NA,
  lab = NA,
  group = NA,
  col = NA,
  title = "Put title here",
  subTitle = "Put subtitle here",
  provider = "Put provider here"
)
}
\arguments{
\item{df}{Dataframe containing the data for the radar chart.}

\item{x}{Variable names (e.g., categories or attributes).}

\item{y}{Percentile values for each category.}

\item{lab}{Data labels to display on the chart - only writes the first group label.}

\item{group}{Grouping variable.}

\item{col}{Series of colors - up to 4 - c("blue","black","orange","red") etc.}

\item{title}{Title for the radar chart.}

\item{subTitle}{Subtitle for the radar chart.}

\item{provider}{Data provider for the data source.}
}
\value{
Returns a radar plot for a single team/player.
}
\description{
Plots a radar chart, up to 4 groups.
}
